refactor(hapi): consolidate api into one file
