+++ lib/hapi/api.js
@@ -1,3 +1,4 @@
+var fs = require('fs')
 var path = require('path')
 var url = require('url')
 
@@ -7,7 +8,10 @@ var Wreck = require('wreck')
 exports.register = register
 exports.register.attributes = {
   name: 'api',
-  dependencies: 'h2o2'
+  dependencies: [
+    'h2o2',
+    'inert'
+  ]
 }
 
 var clientPath = path.dirname(require.resolve('hoodie-client/package.json'))
@@ -25,6 +29,16 @@ function register (server, options, next) {
   })
 
   server.route([{
+    method: 'GET',
+    path: '/{p*}',
+    handler: {
+      directory: {
+        path: options.app.paths.www,
+        listing: false,
+        index: true
+      }
+    }
+  }, {
     method: ['GET', 'PUT', 'POST', 'DELETE', 'OPTIONS'],
     path: '/hoodie/{p*}',
     handler: {
@@ -42,6 +56,16 @@ function register (server, options, next) {
     }
   }, {
     method: 'GET',
+    path: '/hoodie/admin/{p*}',
+    handler: {
+      directory: {
+        path: path.dirname(require.resolve('hoodie-admin-dashboard')),
+        listing: false,
+        index: true
+      }
+    }
+  }, {
+    method: 'GET',
     path: '/hoodie/bundle.js',
     handler: {
       file: path.join(clientPath, 'dist/hoodie.js')
@@ -53,6 +77,23 @@ function register (server, options, next) {
       file: path.join(clientPath, 'dist/hoodie.min.js')
     }
   }])
+
+  // serve app for every 404 on an html page
+  var indexFile = path.join(options.app.paths.www, 'index.html')
+  server.ext('onPostHandler', function (request, reply) {
+    var response = request.response
+
+    if (!response.isBoom) return reply.continue()
+
+    var is404 = response.output.statusCode === 404
+    var isHTML = /text\/html/.test(request.headers.accept)
+
+    // We only care about 404 for html requests...
+    if (!is404 || !isHTML) return reply.continue()
+
+    // Serve index.html
+    reply(fs.createReadStream(indexFile))
+  })
 
   return next()
 }
@@ -83,14 +124,22 @@ exports.addBearerToken = function (err, res, request, reply) {
 }
 
 exports.mapProxyPath = function (db, request, callback) {
+  var headers = request.headers
   // use the bearer token as the cookie AuthSession for couchdb:
-  delete request.headers.cookie
-  if (request.headers.authorization && request.headers.authorization.substring(0, 'Bearer '.length) === 'Bearer ') {
-    request.headers.cookie = 'AuthSession=' + request.headers.authorization.substring('Bearer '.length)
+  delete headers.cookie
+
+  if (headers.authorization && headers.authorization.startsWith('Bearer ')) {
+    headers.cookie = 'AuthSession=' + headers.authorization.substring('Bearer '.length)
   }
 
   // TODO: This is just a temporary fix for PouchDB
-  delete request.headers['accept-encoding']
-  request.headers.host = [db.hostname, db.port].join(':')
-  callback(null, url.resolve(url.format(_.omit(db, 'auth')), request.url.path.substr('/hoodie'.length)), request.headers)
+  delete headers['accept-encoding']
+
+  headers.host = [db.hostname, db.port].join(':')
+  callback(
+    null,
+    url.resolve(url.format(_.omit(db, 'auth')),
+    request.url.path.substr('/hoodie'.length)),
+    headers
+  )
 }
