+++ lib/hapi/index.js
@@ -12,15 +12,11 @@ module.exports = function (env_config) {
 
   return plugins.concat([
     require('./api'),
-    require('./directories'),
-    require('./logger'),
-    require('./handle-404')
+    require('./logger')
   ].map(function (plugin) {
     return {
       register: plugin,
-      options: {
-        app: env_config
-      }
+      options: {app: env_config}
     }
   }), [{
     register: require('hoodie-server-store'),
