+++ lib/hapi/logger.js
@@ -1,8 +1,7 @@
 var qs = require('querystring')
 
-var Squeeze = require('good-squeeze').Squeeze
-
 var log = require('npmlog')
+var Squeeze = require('good-squeeze').Squeeze
 
 exports.register = register
 exports.register.attributes = {
@@ -51,9 +50,7 @@ function register (server, options, next) {
       opsInterval: 30000,
       reporters: [{
         reporter: GoodEvent,
-        config: {
-          callback: logger
-        },
+        config: {callback: logger},
         events: {
           ops: '*',
           response: '*',
